app/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fichas"
version = "0.1.0"
description = "MVP cadastro de fichas de processos - TCM-SP"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "psycopg[binary]>=3.1",
  "pydantic-settings>=2.4",
  "jinja2>=3.1",
  "python-multipart>=0.0.9",
  "passlib[bcrypt]>=1.7",
  "bcrypt<5",
  "itsdangerous>=2.2",
  "python-dotenv>=1.0",
  "google-cloud-storage>=2.16",
  "pytest>=8.0",
  "httpx>=0.27"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
